reporegistry: refactor context menu creation
This doesn't change current functionality, but it will allow
submenus in the repo registry context menu.

Before, a menu was created from a list of action names with all
items at the same level. Now, submenus can be defined by
including a 2-tuple where the first element is the desctipion
text for the submenu root and the second element is a list of
action names to be included.

For example,
['act1', 'act2', ('Submenu', ['act3', 'act4']), 'act5']
creates
- desc for act1
- desc for act2
- Submenu
  - desc for act3
  - desc for act4
- desc for act5