reporegistry: properly add new .hgsub file when adding the first subrepo to a repository
In order to add the first subrepo to a repo we must create an .hgsub file, and
add it to the repository. Up until now, this was not working. We were not able
to add the file to the repo, and we showed an incorrect error message. This
patch fixes both problems.