settings, commit: add tortoisehg.bugtraqtrigger setting
This settings makes some of the SVN issue tracker plugins more compatible with
mercurial's decentralized nature.

Most SVN issue tracker pluggins assume that when you commit you are publising
your changes. Thus they try to update the issue tracker state as soon as you
commit your changes. This is not correct in mercurial's decentralized model,
where committing and publishing are separate steps.

This is also useful when you want to let the issue tracker update the state of
its issues based on the commit messages. Under that model (which is very
common, and is used by Jira and Redmine), you want to be able to get access to
the list of issues, but you do not want to tell the issue tracker to update the
issue state.

This revision adds a new issue.bugtraqtrigger setting to the tortoisehg
section. Only if that setting is set to "commit" will the issue tracker be
updated when clicking the commit button.

Future work could be done to add a "push" mode that would parse the pushed
issues and update them as needed.