commit: do not show the "Issue Tracker" related options on the "Selected Options" field
The issue tracker options are not passed to the commit command. Therefore they
are not very interesting to the user since they do not affect what the commit
will do.

In addition they are not really "selected options", because they are not set
through the "Options" dialog, but through the Settings dialog, which means that
they are not temporary (i.e. they are always applied). Because of this if you
configure an issue tracker you will always get the "Selected Options" field.

Finally, they are often very long because they include the (long) name of the
issue tracker pluggin object. Because they are so long they reduce the file
list size to its minimum size, which makes the refresh button unavailable.

For these reasons it is best not to show those options in the "Selected
Options" field.