phases: handle moving phase backwards and several miscellaneous improvements
The patch series that has been recently pushed was not the most recent version
of that patch series. This revision adds the things that were missing from the
series that was pushed.

The miscellaneus improvements are localization improvements (phase names will
be translated) and an improved way to automatically generate the "Change phase
to..." submenu, so that if new phases are added to mercurial we'll automatically
get them.