qscilib: add "EOL Mode" context submenu and set the EOL automatically
This adds a new submenu to the TortoiseHg editors'context menu which lets the
user select the EOL mode (Windows, Unix or MAC). The default EOL mode is
selected automatically by looking at the EOL of the first line in the editor
text. If there are no EOLs on the editor text (e.g. if the editor is empty) the
default EOL for the current OS is selected instead.

This solves a problem where using the "Rejects" tool could introduce EOL changes
if the EOL of the file being edited was not the same as the system default.

This setting only applies to the current editor.