reporegistry: use a different icon for repositories located on UNC paths
It is usually not a good idea to work with repositories located on remote
drives. It is particularly troublesome to work with repositories located on
UNC paths on windows.

To make it easier to distinguish between "local" and UNC repositories, this
patch modifies the repo registry to use a different icon for repositories
located on UNC paths.

The new icon (remote-repo) is based on a combination of the mercurial-logo.svg
file, available on Wikipedia (http://en.wikipedia.org/wiki/Mercurial) and
modified portions of the "File:Text-html" icon of the Tango icon set
(http://commons.wikimedia.org/wiki/File:Text-html.svg). Both are available under
a Createive Commons License. The icon could certainly be improved a lot but at
least it is different :-)