workbench: add "README" entry to the Help menu
This patch adds a way to setup a "README" file or web address for your
repositories. This is done by adding a new "README" entry to the Help menu.
This new entry is only visible if at least one repository is open and the user
has set the tortoisehg.readme configuration key.

When tortoisehg.readme has been set the new menu entry will appear on the
help menu. When clicked the file or URL specified on that configuration key will
be open.

This is useful for people who want to customize TortoiseHg for their needs, and
want to supply additional documentation. In conjunction with the Projrc
extension it is possible to set per-repository help files as well.