workbench: add GitHub/BitBucket-like "README" file detection
In addition to being able to configure a README file or URL by setting the
tortoisehg.readme configuration key, let TortoiseHg also look for "README" files
on the current repo's root folder.

This is similar to what github and bitbucket do on their repositories.

With this patch, README files can be set in 3 ways, which are checked in the
following order of decreasing priority:

1.- From the tortoisehg.readme key on the current repo's configuration file
2.- An existing "README" file found on the repository root
        * Valid README files are those whose name matches one of the following
          (in a non case-sensitive way):
                readme.txt, read.me, readme.html, readme.pdf,
                readme.doc, readme.docx, readme.ppt, readme.pptx,
                readme.md, readme.markdown, readme.mkdn, readme.rst,
                readme.textile, readme.rdoc, readme.asciido, readme.org,
                readme.mediawiki, readme.creole,readme.pod, readme
        * Note that these names are a combination of those supported by GitHub
          and BitBucket.

3.- From the tortoisehg.readme key on the user's global configuration file.