sync: improve pull/push behavior when no remote path is set
The original error message was "No URL selected", which was confusing (see
http://coderscentral.blogspot.com/2011/07/how-to-have-fanatical-user-base.html
for an example of several frustrated new users).

This patch improves the situation in two ways:
- Improves the error message
- Automatically shows the Sync widget to make it easier for new users to
discover it.

In order to show the sync widget I have added a new signal to the SyncWidget
object, called "switchToRequest", defined as:

switchToRequest = pyqtSignal(QString)

Then I have connected this signal to the repowidget switchToNamedTaskTab()
method in the createSyncWidget method as follows:

sw.switchToRequest.connect(self.switchToNamedTaskTab)

Since switchToNamedTaskTab() expects to receive a Python string, it now needs
to convert its input into a python string.