commit, bugtraq: delay the bugtraq button setup to speed up the commit widget creation (closes #1734)
Creating the bugtraq button on the Commit widget takes time because we must
create an IBugTraqProvider2 COM object. On a pretty powerful machine takes
0.4 seconds or less (and it can take several seconds on a less powerful machine).

The solution is to create a "disabled" placeholder button, and setup the button
(with the corresonding COM object creation) after the commit widget has been
created (100 ms), at which point the button can be enabled.

The result is that the commit widget is created as fast as if no issue tracker
was configured.