qtlib: add ChoicePrompt dialog class
This new class shows a dialog containing a message, a combobox and an OK and
Cancel button.

When the user clicks OK the dialog returns the selected item in the combo box.
If the user clicks Cancel or closes the dialog the dialog returns None.

In order to actually show the dialog the run() method must be called.