workbench, reporegistry: show progress dialog for lengthy startup operations
Loading the workbench may take a long time, particularly if you have
repositories located on remote drives. Up until now, the user did not get any
feedback or visual cue while the workbench is being loaded . The user cannot
even tell that the workbench is being loaded in the first place!

With these changes a progress dialog will be shown when the subrepos are being
loaded by the reporegistry and while the workbench is reopening the repos that
where open on the last workbench session.

This informs the user that the workbench is being loaded.

In addition, once the workbench is loaded, reloading the repo-registry will show
a progress bar on the status bar.