qscilib: add "TAB inserts" context submenu
This adds a new submenu to the TortoiseHg editors'context menu which lets the
user select whether hitting TAB inserts a TAB or spaces.

As with the "EOL Mode" menu, this new submenu includes an "Auto" mode, which is
the default, which will try to detect whether the current file uses TABS or
spaces for indentation. This is done by checking the first character on the
first 100 lines. The first line that starts with a space or a TAB determines
whether the editor will use spaces or TABS for indenting.