docklog: raise output log when shown
Up until now clicking on the "Show log" link that is shown on the infobar when
there is an error does not bring the log panel to the front. It does only ensure
that the log panel is visible. If the log panel is visible but is docked with
other panels and is not the topmost docked panel, clicking on the show log
link wdoes nothing.

Clicking on the Show Log button or in the View menu does not bring the log
output to the front either.

This revision fixes this problem. As a side effect, whenever there is an error
the output log is brought to the front as well, which I think is a good idea.

This change could be applied to other panels as well (perhaps by creating a
common base class, based upon the QDockWidget, which all such panels could
inherit from.