manifest: do not try to edit folders except on the working directory
Up until now you could try to call the "edit local" command on a folder,
which would result on an obscure error message.

This patch fixes that. It will still let you call "edit" on folder when
viewing the working directory, in which case it will "edit" all the local files
in that folder.

This may increase the time that it takes to edit a file on big repositories,
since we must check whether the requested file is in the repository.