manifest: handle revision links in dialog mode
This makes it possible to click on the revision links on the revision details
on the manifest dialog to update the dialog to the selected revision.

Up until now a "linkActivated" signal would be emitted, but if the manifest
dialog was open in stand alone mode there was no one to catch it.

It seems that the manifest dialog is never opened from the workbench, which
means that it is ok to handle the link right on the manifest dialog.