manifest: show file icons according to the file type
Rather than showing a generic "file icon" on the manifest, show the icon that
corresponds to each file type.

This seems to also speed up the manifest dialog, particularly when it comes to
filtering the manifest file list contents. The speed up is due to the catching
of the icon object for each tree element, whereas before it was being constantly
recreated.