email: add support for new hg email --body flag
A new "body" checkbox has been added to the email dialog. When enabled, the
patch contents will be included on the the body of the message.

The "body"checkbox" will be automatically enabled if neither the "attach" nor
the "inline" checkboxes are enabled.

If the mercurial version is 2.1 or lower, checking any of the attachment
options (attach or inline) will automatically uncheck the body checkbox.
Disabling both checkboxes will automatically check the body checkbox.