reporegistry: improve accuracy of subrepo detection when adding a repository (refs #1672)
Up until now adding a subrepo that was contained inside another subrepo would
always mark the inner repo as a subrepo. This is not always true.

With this patch, we also check whether the added subrepo is _currently_ a
subrepo of the outer repository. We respect the "showNetworkSubrepos" flag
(i.e. we'll only check if an added repo is a subrepo if the repo is local
or if the showNetworkSubrepos flag is set).

Note that this does not a full fix issue #1672, because we only perform this
check once, when adding the repo, and because we only check whether the repo
is _currently_ a subrepo of the outer repo. However I think that it is an
improvement over the old behavior.