settings: add 'Workbench/Activate Bookmarks' setting
This new setting allows configuring when will TortoiseHg show a prompt to
activate a bookmark that is found on the revision that is being updated to.

There are three options:

- auto: Try to automatically activate bookmarks. When updating to a revision
that has a single bookmark it will be activated automatically. Show a prompt if
there is more than one bookmark on the revision that is being updated to.
- prompt: The default. Show a prompt when updating to a revision that has one or
more bookmarks.
- never: Never show any prompt to activate any bookmarks.