workbench: improve push infobar message
Up until now the infobar message that is shown on outgoing had a button with
the text 'Push' on it. This did not take into account the
'tortoisehg.defaultpush' setting, which can change the behavior of the push
button.

This patch shows a different text on te push button depending on the defaultpush
setting ('Push', 'Push current branch' or 'Push current revision').

I thought about putting the actual branch name or revision number on the button
as well. However this could lead to error if the user updates to a different
revision or branch while the infobar push message is being shown.

I could fix the rev and/or revision that is being pushed to when you click the
button, rather than using the current one to solve that issue. However I think
that is less flexible than the current solution (that let's you update to a
different revision before push without having to do a new update), and it could
also break down if the user stripped the current revision before hitting push.

My only concern is that an unexperienced user may believe that 'Push current
revision' will only push the current revision, not its ancestors. However this
should not be a common mistake (and if it is, the user has greater problems
anyway :-P). Another concern is that a user could believe that 'current revision'
equals 'selected revision'. In any case I still think that this is an
improvement over the current situation, where you do not know what you are
pushing.