import: show the Import button when import fails
Currently when you click the Import button it is hidden. If the import fails
the Import button remains hidden. The only way to try to import again is to
close the import dialog and reopen it, which means that you must re-select the
patches that you want to import.

With this patch you'll be able to try to import again after an import failure.
If you try to import again the import will fail again. However you'll be able
to change the import target so that you can import into a patch, for example.