repowidget: remove 'Browse at revision...' from context menu
The existing 'Browse at revision...' entry on the revision context menu is
redundant since you can already open the manifest by clicking on the manifest
button on the widget toolbar.

Removing this command simplifies the revision context menu a little, which I
think is a good idea, since it is getting a little too complex.