repowidget: add support for matching multiple revisions at a time
This makes it possible to call the "Find similar revisions" dialog with a set
of revisions, rather than with a single revision.

It was already possible to type a revision set on the match dialog, but it was
not possible to select multiple revisions on the repowidget and open the match
dialog with those revisions selected.