commit: show the Amend command on the commit button even if MQ is disabled
With this patch we do not create a Commit button and an MQMenu anymore. Instead
we create a single CommitButtonMenu, which shows different options depending on
whether the MQ extension is enabled or not.

This patch renames all the classes, methods, properties and functions related
to the "MQMenu", using "commit" or "commitbutton" or "commitb" where "mq" was
used before.

The meaning of the "hasmqbutton" property of the commit widget has changed. Now
it is True if the "CommitButtonMenu" has MQ entries (i.e. if the mq extension is
enabled). Perhaps its name could be changed into "hasmqextension" or
"mqisenabled".