workbench: make the task toolbar buttons configurable
This adds a new configuration key: tortoisehg/workbench.task-toolbar which
can be set to choose which buttons are shown on the task toolbar and to select
their order.

This key must contain a list of task names separated by one or more spaces.
Valid task names are: commit, mq, pbranch, log, manifest, grep and sync.

Note that the mq and pbranch buttons are only shown if the corresponding
extensions are enabled, regardless of the value of this configuration key.

If this key is not set the original task buttons will be shown in their original
order, as if the key had been set to: log commmit mq sync manifest grep pbranch

It is possible to add separators between task buttons by add any string that
does not match one of the valid task names (e.g. | or sep or -, etc)

For example:

[tortoisehg]
workbench.task-toolbar = commit mq pbranch | log  manifest | grep | sync

would be equivalent to:

[tortoisehg]
workbench.task-toolbar = commit mq pbranch - log  manifest separator grep | sync

I believe that the current default ordering could be improved, but that is a
matter for another patch.