diff: rename 'Visual diff' to 'Diff to parent'
In many context menus there are two 'diff' entries:
'Visual diff' and 'Diff to local'.

Both do basically the same, except that one diffs to one of the revision parents
while the other diffs to the working directory.

This change makes the context menu entries more consistent and it also makes it
easier to tell which does which.

I have also made the capitalization of these entries more uniform, as well as
removed 'Visual' from some 'Visual Diff to Local' entries to make
the menues more consistent.