workbench: clear previous session open repo list and last active respo on startup
This patch clears the "openrepos" list (and the "lastactiverepo" value) when
the previous session is restored. This ensures that opening a new workbench
window when there is already an existing workbench window will not reopen the
repos that were open on the previous session again. Instead a new, empty
workbench window will be open.

Note that if the first workbench window was open with a "root" path, that
root path will be open on the new workbench window (this was already the case
without this patch).