commit: automatically set 'close branch' commit message
If the commit message is empty, when the user opens the "branchop" dialog and
selects "Close current branch", automatically set the commit message to:

Close BRANCH_NAME branch

where BRANCH_NAME is the current branch name.
The 'tortoisehg.engmsg' setting is respected. If set to False (or not set) the
commit message will be translated. Otherwise the message will be in English.