fileview: add 'Annotate Options' context submenu (fixes #1861)
Add a new submenu ('Annotate Options') to the annotate view context menu.
This new submenu lets you chose which info should be shown on the Annotate View
margin.

Valid values options are:
- Show Author
- Show Date
- Show Revision

The user chan check one or more of these options. At least one of them must be
enabled at any given time.

The selections on this new context menu are persistent, since they are saved as
a QSetting.

Independently of the mode that is used, this change has the drawback that it is
a bit slow. It introduces a delay between the moment the file is displayed and
the moment the annotations are shown.

The code that calculates the annotation margin caches the results for each
revision that it encounters. There may be additional optimizations that could
be performed to improve the speed of this code further.