filectxactions: add "Folder history" command
When the folder specific manifest context menu was introduced the "File history"
and "Compare revisions" commands were removed for folders, because they did not
work in the first place.

This patch introduces a "Folder history" command which is somewhat similar to
the "File history" command.

Rather than opening a new dialog showing the file revisions, this command simply
sets the repowidget filter to only show revisions where any of the files
contained in the selected folder were modified.