filectxactions: add folder specific context menu
Up until now folders and files shared the same context menu on the manifest
widget. There are several commands which make sense for files but not for
folders and vice-versa.

This patch adds a separate context menu for folders, which has some different
entries than the regular file context menu. In doing so it also renames some
actions which are shared between subrepos and folders.

This patch removes the 'ledit' action for folders. Although it does work I don't
think it is very useful or makes much sense (it opens all files contained in the
folder on the selected editor).