manifest: expand the file tree when "return" is pressed
This makes it possible to expand the file tree after filtering. This is a less
ambitious patch than my recent series that would autoexpand the tree by default.

While I still hope to get that series into 2.5, we are still debating a few GUI
issues. While those are resolved I think that this is a good step forward.