outgoing: fix push all command and improve outgoing infobar
The "Push All" command was broken. Calling it from the repowidget context menu
or from the "Push" outgoing infobar button would not always push all. Instead,
it would use the tortoisehg.defaultpush setting.

This patch fixes that, and in doing so it improves the outgoing infobar button
to explicitly specify the revision or branch that will be pushed. It also fixes
a possible issue when trying to push revision 0.