filectxactions: make 'Open subrepo' select the selected subrepo revision
Up until now using 'Open subrepo' would simply open the subrepo in a new (or
existing) tab. However the selected revision on the subrepo would not be the
revision in which the subrepo was set for the given parent repo revision.

This patch changes that. For example, if subrepo 's' was at revision 3 in its
parent repo revision 7, using 'Open subrepo' on the subrepo 's' on its
parent repo revision 7 will open the subrepo s and then automatically select
its revision number 3.

To implement this I've changed the way the "linkActivated" signal is emitted for
subrepos. I've added the revision hash after the subrepo 'URI'. To do so I've
used the usual URI convention, where a '?' separates the actual URI path from
its parameters. Obviously this assumes that no subrepo path will have a '?'
character on it, which is true at least on Windows, and probably quite unlikely
on other platforms.

Note that the behavior when selecting "Open subrepo" fromt he commit widget has
not been changed.