reporegistry: ensure that the selected tab item is visible
Call the RepoTreeView.scrollTo() method to ensure that the repo registry item
corresponding to the selected repo tab is visible.

I'm not super happy with the name of the auxiliary function _getItemAndAncestors,
but I could not find a better name and I did not want to repeate that code in
expand() and scrollTo(). There is also a bit of repeated code in those functions
but I could not find a nice way to share that code.