workbench: make the 'Single workbench mode' the default
Most programs that support 'tabs', such as Firefox, Internet Explorer, Eclipse,
etc. try to reuse their existing windows by default when opening new 'tab'. In
order to bring the behavior of TortoiseHg in line with this typical behavior,
the 'Single workbench mode' must be the default.

This was not possible to do until recently, because the 'single workbench mode'
still had some quirks, such as not bringing the workbench to the front when a
new workbench was open from the explorer context menu. Now that these quirks
have been fixed, I think it is time to make the 'single workbench mode' the
default.

This has the added benefit that the 'single workbench mode' handles sessions
more consistently, because it properly remembers all tabs that were open when
the single workbench window was last closed.

Note that in this revision I also fixed a small, related typo.