revdetails, manifest: update filectxactions when a file is clicked
The filectxactions were being updated only on menurequest. The result was that
using any of the file context menu keyboard shortcuts on the revdetails or
manifest file list would fail if the context menu had never been shown, or
if the context menu had been shown the actions would be triggered on the wrong
file (on the file for which the menu had been shown).

This is only a partial fix for this problem. The problem can also happen if the
user uses the keyboard to navigate the file list, in which case the
filectxactions will not be updated. I dont think that the revdetails filelist
nor the manifest treeview have a signal that is triggered when the selection
changes, so this is not as easy to fix.