workbench: update the custom toolbar when the repo tab changes
Up until now the custom toolbar was populated once, when the workbench was
created, and it would show the global custom commands plus the custom commands
of the the startup repo (if any). With this change the custom toolbar is updated
whenever the repo tab changes.

Note that we could probably "cache" the toolbars rather than generating them
again. The advantage of generating them again is that this lets us pick up tool
configuration changes on the fly, although we could do this as well if we
invalidated the cache when the config changes. In addition, in regular usage it
seems that once you generate the toolbar once, regenerating it is quite fast.