tortoisehgtools: change tool definition format
The way in which "custom tools" are defined is changed in a non backward
compatible way with this patch. Rather than using a "location" field to select
in where the tool is shown on the GUI we now there is a configuration key for
each GUI location. Each of those configuration keys must be set to a list of
tool names, which are the ones that will be shown on the corresponding GUI
"location". You can also place a "|" character to indicate that a separator
should be shown.

In addition, it is now possible to limit the tools that the
hglib.tortoisehgtools() function returns, by specifying the "location" of the
tools that must be shown, as well as the "sources" (i.e. the hgrc files) that
must be read. Finally, it is possible to pass a wconfig object rather than an
ui config to that function.