docklog: add command history (closes #1779)
It is now possible to use the up and down arrows to perform an incremental
search (backwards or forward) through the previous commands that were typed into
the output log.

In addition:
- Pressing "TAB" shows the commands whose beginning matches the
currently typed command text.
- Pressing "ESC" while the cursor is at the command prompt clears the current
command prompt text.
- Pressing "ESC" while the cursor is not at the command prompt moves the cursor
to the command prompt, keeping the command prompt text if any.