docklog: add 'hg/thg command' completion
If the user types a mercurial or tortoisehg command (i.e. starting with 'hg' or
'thg') and presses TAB in addition to showing matching history items, show
relevant mercurial/tortoisehg commands and/or command options.

If both matching history commands and mercurial/thg commands are shown, the
matching mercurial/thg commands are shown after the history commands.

Also, do not show the list of matching commands if there is only a single
matching command (just complete it).