qscilib: add icons to 'Next' and 'Prev' search buttons
The icons are the 'go-down' and 'go-up' icons respectivelly.
These are the same icons that are used on the fileview to go to the next or
previous difference. However I think that the context (i.e. the fact that the
buttons are on the search toolbar) make it obvious that these buttons do not do
the same as the existing 'next diff' and 'previous diff' buttons.

Note that the existing '>>' and '<<' parts of the text labels have been removed,
since they are not necessary anymore.