repowidget: add graft support
Adds a new "graft dialog" that can be shown by using a new "Graft" command that
is shown on the revision context menu, where the transplant command usually is.

The graft dialog is based on the rebase dialog. I have hopefully removed all
unnecessary code from that dialog.

Through this dialog it is possible to resolve any conflicts that may happen
during a graft, and it is also possible to abort or continue an ongoing graft.

When grafting multiple revisions, the dialog shows the list of revisions that
will be grafted, in their proper order as well as the information regarding the
revision that is being currently grafted.

If the user tries to graft a different set of revisions while an existing graft
operation is incomplete, a warning will be shown, and the user will be able to
either cancel or show the graft dialog to continue with or abort _the original_
graft operation, not new one.

The only option that the dialog lets you change is whether conflicts should be
automatically resolved. It is not possible to set the merge tool which I think
is unnecessary, since the resolve dialog will let you select a tool anyway.

I think it would be best to move all the "history manipulation" commands (such
as graft and phase changes) into the "Modify history" submenu, but that should
be done on another patch.

The dialog uses the transplant icon but this icon does not represent graft well.
I think a new icon should be made, perhaps one based on the rebase icon (without
the red cross on the source revision).