repowidget, status, manifest: change background color when a filter is applied
This patch changes the background color of filtered lists and trees.

This takes into account whether the default background color is light or dark.
If the default background color is light, the filtered color is a light yellow,
while when the background color dark, the filtered color is a dark grey.

In order to make it easy to filter different lists with the same colors, a new
PaletteSwitcher class has been added to qtlib. An instance of this class must
be added to any view that must change its background color when filtered. The
class constructor takes the "parent" or "target" object as its only parameter.

In order to change the background color, one must simply call the
"enablefilterpalette" method of the PaletteSwitcher instance to enable or
disable the background change.