shortcuts: give Ctrl+Shift+L for openlocal (fixes #2118)
Ctrl+Shift+O conflicts with "Show Repository Registry".

This unifies the shortcut for openlocal (=openfile), and removes the shortcut
from FilectxActions.editlocal, which was previously Ctrl+Shift+L.

In manifest view, Ctrl+Shift+E, which is the shortcut for WctxActions.edit,
is assigned not to editlocal, but to editfile action. I guess that's because
the manifest view is intended to handle committed (non-local) data.