filedialogs: add multi-selection menu (fixes #1871)
This change backs out bcf0854d848c6534 and adds a new menu which is shown when
more than one item is selected. The menu shows two options:

- Diff selected changesets...
- Diff selected file revisions...

Ideally it should not be possible to select more than two revisions. However,
for now if more than two revisions are selected and the user right clicks the
context menu will not be shown and insted a message will be shown on the status
bar indicating that you cannot select more than two revisions.

The code to generate the menus could be improved a lot, but that should be left
for another revision.