commit: prompt the user to set a default username on commit if necessary
Currently TortoiseHg does not ever prompt the user to set its default username.
This means that is quite easy to make a commit with an undesired username. This
is particularly true on Windows, where the default username is the current
account name followed by "@" and the full network computer name.

With this change TortoiseHg checks whether a default username is set, and if not
it prompts the user to set one (letting the user open the settings dialog right
from the prompt).